[xds_package_non_experimental] xds: move balancer/resolver registation to package xds
Instead of package "xds/experimental".
