+++ examples/features/xds/client/main.go
@@ -31,6 +31,7 @@ import (
 	"google.golang.org/grpc"
 	pb "google.golang.org/grpc/examples/helloworld/helloworld"
 
+	// TODO: change this to xds after the xds package is released.
 	_ "google.golang.org/grpc/xds/experimental" // To install the xds resolvers and balancers.
 )
